+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Make imenu recognize indented def-forms
 * New interactive command `clojure-cycle-when`.
 * New interactive command `clojure-cycle-not`.
+* New defcustom `clojure-comment-regexp` for font-locking `#_` or `#_` AND `(comment)` sexps.
 
 ## 5.6.1 (2016-12-21)